added list all books and show one

+++ src/main/java/parkanthony/services/BookService.java
@@ -32,8 +32,18 @@ public class BookService {
             return null;
         }
     }
-
+    public Book update(Long id, String title, String desc, String lang, int pages) {
+    	Book edited = bookRepository.findById(id).get(); // repo find -> gets Optional and .get returns value (book instance or row)
+    	edited.setTitle(title);
+    	edited.setDescription(desc);
+    	edited.setLanguage(lang);
+    	edited.setNumberOfPages(pages); // we make edits
+    	return bookRepository.save(edited); // use (dot)save and puts book row back into Repo data
+    }
     public void deleteBook(Long id) {
-    	bookRepository.deleteById(id);
+    	Optional<Book> optionalBook = bookRepository.findById(id);
+        if(optionalBook.isPresent()) {
+            bookRepository.deleteById(id);
+        }
     }
 }
